enum Error and strut ActorRef has been moved to src/actor_ref.rs and src/error.rs respectively.
