+++ src/error.rs
@@ -1,0 +1,101 @@
+// Copyright 2022 Jeff Kim <hiking90@gmail.com>
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::Identity;
+use std::time::Duration;
+
+#[derive(Debug)]
+/// Represents errors that can occur in the rsactor framework.
+pub enum Error {
+    /// Error when sending a message to an actor
+    Send {
+        /// ID of the actor that failed to receive the message
+        identity: Identity,
+        /// Additional context about the error
+        details: String,
+    },
+    /// Error when receiving a response from an actor
+    Receive {
+        /// ID of the actor that failed to send a response
+        identity: Identity,
+        /// Additional context about the error
+        details: String,
+    },
+    /// Error when a request times out
+    Timeout {
+        /// ID of the actor that timed out
+        identity: Identity,
+        /// The duration after which the request timed out
+        timeout: Duration,
+        /// Type of operation that timed out (e.g., "send", "ask")
+        operation: String,
+    },
+    /// Error when a message type is not handled by an actor
+    UnhandledMessageType {
+        /// ID of the actor that timed out
+        identity: Identity,
+        /// Expected message types
+        expected_types: Vec<String>,
+        /// Actual message type ID
+        actual_type_id: std::any::TypeId,
+    },
+    /// Error when downcasting a reply to the expected type
+    Downcast {
+        /// ID of the actor that sent the incompatible reply
+        identity: Identity,
+        expected_type: String,
+    },
+    /// Error when a runtime operation fails
+    Runtime {
+        /// ID of the actor where the runtime error occurred
+        identity: Identity,
+        /// Additional context about the error
+        details: String,
+    },
+    /// Error related to mailbox capacity configuration
+    MailboxCapacity {
+        /// The error message
+        message: String,
+    },
+}
+
+impl std::fmt::Display for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Error::Send { identity: actor_id, details } => {
+                write!(f, "Failed to send message to actor {}: {}", actor_id.name(), details)
+            }
+            Error::Receive { identity: actor_id, details } => {
+                write!(f, "Failed to receive reply from actor {}: {}", actor_id.name(), details)
+            }
+            Error::Timeout { identity: actor_id, timeout, operation } => {
+                write!(f, "{} operation to actor {} timed out after {:?}",
+                       operation, actor_id.name(), timeout)
+            }
+            Error::UnhandledMessageType { identity: actor_id, expected_types, actual_type_id } => {
+                write!(
+                    f,
+                    "Actor '{}' received an unhandled message type. Expected one of: [{}]. Actual message type ID: {:?}",
+                    actor_id.name(),
+                    expected_types.join(", "),
+                    actual_type_id
+                )
+            }
+            Error::Downcast { identity: actor_id, expected_type } => {
+                write!(f, "Failed to downcast reply from actor {} to expected type '{}'",
+                    actor_id.name(), expected_type)
+            }
+            Error::Runtime { identity: actor_id, details } => {
+                write!(f, "Runtime error in actor {}: {}", actor_id.name(), details)
+            }
+            Error::MailboxCapacity { message } => {
+                write!(f, "Mailbox capacity error: {}", message)
+            }
+        }
+    }
+}
+
+impl std::error::Error for Error {}
+
+/// A Result type specialized for rsactor operations.
+pub type Result<T> = std::result::Result<T, Error>;
